Upgrade to the new SDoc syntax: [SECTION] -> [[SECTION]]
As of the version 0.9.1, StrictDoc has switched to a new syntax for
defining SECTIONs. From now on, all sections must be defined as

`[[SECTION]]` instead of `[SECTION]`.

The rationale for the migration is explained in the related GitHub
ticket:

[Migration: SDoc model: Merge Section, Node, Composite Node into just
Node #2193](https://github.com/strictdoc-project/strictdoc/issues/2193)

**IMPORTANT:** When this PR is reviewed and merged, all existing PRs
should be updated to only have [[SECTION]] tags instead of [SECTION].

See also the migration guide for further reference:

https://strictdoc.readthedocs.io/en/stable/stable/docs/strictdoc_01_user_guide.html#SECTION-UG-NODE-MIGRATION

**StrictDoc 0.9.1 version is needed for this changeset to work.**

Signed-off-by: Stanislav Pankevich s.pankevich@gmail.com